+++ src/main/java/kg/peaksoft/peaksoftlmsab4/dto/admin/AdminResponse.java
@@ -1,16 +1,14 @@
-package kg.peaksoft.peaksoftlmsab4.dto.response;
+package kg.peaksoft.peaksoftlmsab4.dto.admin;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class InstructorResponse {
+public class AdminResponse {
+
     private String id;
     private String firstName;
     private String lastName;
-    private String mobilePhone;
-    private String specialization;
     private String email;
 }
-
