+++ src/main/java/kg/peaksoft/peaksoftlmsab4/dto/auth/AuthRequest.java
@@ -1,4 +1,4 @@
-package kg.peaksoft.peaksoftlmsab4.dto.request;
+package kg.peaksoft.peaksoftlmsab4.dto.auth;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,11 +7,7 @@ import javax.validation.constraints.NotBlank;
 
 @Getter
 @Setter
-public class AdminRequest {
-    @NotBlank
-    private String firstName;
-    @NotBlank
-    private String lastName;
+public class AuthRequest {
     @NotBlank
     private String email;
     @NotBlank
