--- src/main/java/kg/peaksoft/peaksoftlmsab4/dto/request/AuthRequest.java
@@ -1,15 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@Setter
-public class AuthRequest {
-    @NotBlank
-    private String email;
-    @NotBlank
-    private String password;
-}
