--- src/main/java/kg/peaksoft/peaksoftlmsab4/dto/response/AdminResponse.java
@@ -1,14 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.dto.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AdminResponse {
-
-    private String id;
-    private String firstName;
-    private String lastName;
-    private String email;
-}
