--- src/main/java/kg/peaksoft/peaksoftlmsab4/dto/response/StudentResponse.java
@@ -1,16 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.dto.response;
-
-import kg.peaksoft.peaksoftlmsab4.model.enums.StudyFormat;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class StudentResponse {
-
-    private String id;
-    private String firstName;
-    private String mobilePhone;
-    private StudyFormat studyFormat;
-    private String email;
-}
