--- src/main/java/kg/peaksoft/peaksoftlmsab4/exception/BadRequestException.java
@@ -1,12 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.exception;
-
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
-
-@ResponseStatus(BAD_REQUEST)
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
