+++ src/main/java/kg/peaksoft/peaksoftlmsab4/exception/NotFoundException.java
@@ -1,12 +1,8 @@
 package kg.peaksoft.peaksoftlmsab4.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
 public class NotFoundException extends RuntimeException{
-
     public NotFoundException() {
+        super();
     }
 
     public NotFoundException(String message) {
