--- src/main/java/kg/peaksoft/peaksoftlmsab4/model/enums/StudyFormat.java
@@ -1,6 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.model.enums;
-
-public enum StudyFormat {
-    ONLINE,
-    OFFLINE
-}
