--- src/main/java/kg/peaksoft/peaksoftlmsab4/repository/InstructorRepository.java
@@ -1,9 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.repository;
-
-import kg.peaksoft.peaksoftlmsab4.model.entity.Instructor;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface InstructorRepository extends JpaRepository<Instructor,Long> {
-}
