--- src/main/java/kg/peaksoft/peaksoftlmsab4/repository/StudentRepository.java
@@ -1,9 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.repository;
-
-import kg.peaksoft.peaksoftlmsab4.model.entity.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StudentRepository extends JpaRepository<Student,Long> {
-}
