+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/AuthService.java
@@ -1,0 +1,8 @@
+package kg.peaksoft.peaksoftlmsab4.service;
+
+import kg.peaksoft.peaksoftlmsab4.dto.auth.AuthRequest;
+import kg.peaksoft.peaksoftlmsab4.dto.auth.AuthResponse;
+
+public interface AuthService {
+    AuthResponse authenticate(AuthRequest authRequest);
+}
