--- src/main/java/kg/peaksoft/peaksoftlmsab4/service/StudentService.java
@@ -1,4 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.service;
-
-public interface StudentService {
-}
