--- src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/StudentServiceImpl.java
@@ -1,7 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.service.serviceImpl;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class StudentServiceImpl {
-}
