+++ src/test/java/kg/peaksoft/peaksoftlmsab4/PeaksoftlmsAB4ApplicationTests.java
@@ -6,8 +6,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class PeaksoftlmsAB4ApplicationTests {
 
+    private final
+
     @Test
     void contextLoads() {
+
     }
 
 }
